- Fix Norm

+++ include/link_func_t.h
@@ -20,8 +20,8 @@ typedef struct link_func_s
 
 typedef struct data_link_s
 {
-	ssize_t start;
-	ssize_t end;
+	unsigned long long start;
+	unsigned long long end;
 	char *name;
 } data_link_t;
